refactor: inline rxdb.ts into rxdb-adapter
- Move plugin setup + storage config directly into adapter
- Delete shallow wrapper at src/lib/rxdb.ts
