--- src/lib/rxdb.ts
@@ -1,36 +1,0 @@
-import { createRxDatabase, addRxPlugin, type RxJsonSchema, type RxDatabase } from 'rxdb';
-import { RxDBDevModePlugin } from 'rxdb/plugins/dev-mode';
-import { RxDBQueryBuilderPlugin } from 'rxdb/plugins/query-builder';
-import { getRxStorageDexie } from 'rxdb/plugins/storage-dexie';
-import { RxDBUpdatePlugin } from 'rxdb/plugins/update';
-import { wrappedValidateAjvStorage } from 'rxdb/plugins/validate-ajv';
-
-if (import.meta.env.DEV) {
-	addRxPlugin(RxDBDevModePlugin);
-}
-
-addRxPlugin(RxDBUpdatePlugin);
-addRxPlugin(RxDBQueryBuilderPlugin);
-
-const storage = wrappedValidateAjvStorage({
-	storage: getRxStorageDexie()
-});
-
-export const createDatabase = (name: string) =>
-	createRxDatabase({
-		name,
-		storage,
-		ignoreDuplicate: import.meta.env.DEV
-	});
-
-export const createCollection = async <T extends object>(
-	db: RxDatabase,
-	collectionName: string,
-	schema: RxJsonSchema<T>
-) => {
-	await db.addCollections({
-		[collectionName]: {
-			schema
-		}
-	});
-};
